refactor(structure): move all generators to a common module, and rename
all names containing `ExampleGenerator` are deprecated; instead, just use `Generator`
