+++ tests/test_ode.py
@@ -8,7 +8,8 @@ matplotlib.use('Agg') # use a non-GUI backend, so plots are not shown during tes
 from neurodiffeq import diff
 from neurodiffeq.networks import FCNN, SinActv
 from neurodiffeq.ode import IVP, DirichletBVP
-from neurodiffeq.ode import solve, solve_system, Monitor, ExampleGenerator
+from neurodiffeq.ode import solve, solve_system, Monitor
+from neurodiffeq.generator import Generator1D
 
 import torch
 torch.manual_seed(42)
@@ -29,44 +30,44 @@ def test_train_generator():
     exponential = lambda x, t: diff(x, t) - x
     init_val_ex = IVP(t_0=0.0, x_0=1.0)
 
-    train_gen = ExampleGenerator(size=32, t_min=0.0, t_max=2.0, method='uniform')
+    train_gen = Generator1D(size=32, t_min=0.0, t_max=2.0, method='uniform')
     solution_ex, _ = solve(ode=exponential, condition=init_val_ex,
                            t_min=0.0, t_max=2.0,
                            train_generator=train_gen,
                            max_epochs=3)
-    train_gen = ExampleGenerator(size=32, t_min=0.0, t_max=2.0, method='equally-spaced')
+    train_gen = Generator1D(size=32, t_min=0.0, t_max=2.0, method='equally-spaced')
     solution_ex, _ = solve(ode=exponential, condition=init_val_ex,
                            t_min=0.0, t_max=2.0,
                            train_generator=train_gen,
                            max_epochs=3)
-    train_gen = ExampleGenerator(size=32, t_min=0.0, t_max=2.0, method='equally-spaced-noisy')
+    train_gen = Generator1D(size=32, t_min=0.0, t_max=2.0, method='equally-spaced-noisy')
     solution_ex, _ = solve(ode=exponential, condition=init_val_ex,
                            t_min=0.0, t_max=2.0,
                            train_generator=train_gen,
                            max_epochs=3)
-    train_gen = ExampleGenerator(size=32, t_min=0.0, t_max=2.0, method='equally-spaced-noisy', noise_std=0.01)
+    train_gen = Generator1D(size=32, t_min=0.0, t_max=2.0, method='equally-spaced-noisy', noise_std=0.01)
     solution_ex, _ = solve(ode=exponential, condition=init_val_ex,
                            t_min=0.0, t_max=2.0,
                            train_generator=train_gen,
                            max_epochs=3)
-    train_gen = ExampleGenerator(size=32, t_min=np.log10(0.1), t_max=np.log10(2.0), method='log-spaced')
+    train_gen = Generator1D(size=32, t_min=np.log10(0.1), t_max=np.log10(2.0), method='log-spaced')
     solution_ex, _ = solve(ode=exponential, condition=init_val_ex,
                            t_min=0.1, t_max=2.0,
                            train_generator=train_gen,
                            max_epochs=3)
-    train_gen = ExampleGenerator(size=32, t_min=np.log10(0.1), t_max=np.log10(2.0), method='log-spaced-noisy')
+    train_gen = Generator1D(size=32, t_min=np.log10(0.1), t_max=np.log10(2.0), method='log-spaced-noisy')
     solution_ex, _ = solve(ode=exponential, condition=init_val_ex,
                            t_min=0.1, t_max=2.0,
                            train_generator=train_gen,
                            max_epochs=3)
-    train_gen = ExampleGenerator(size=32, t_min=np.log10(0.1), t_max=np.log10(2.0), method='log-spaced-noisy', noise_std=0.01)
+    train_gen = Generator1D(size=32, t_min=np.log10(0.1), t_max=np.log10(2.0), method='log-spaced-noisy', noise_std=0.01)
     solution_ex, _ = solve(ode=exponential, condition=init_val_ex,
                            t_min=0.1, t_max=2.0,
                            train_generator=train_gen,
                            max_epochs=3)
 
     with raises(ValueError):
-        train_gen = ExampleGenerator(size=32, t_min=0.0, t_max=2.0, method='magic')
+        train_gen = Generator1D(size=32, t_min=0.0, t_max=2.0, method='magic')
     print('ExampleGenerator test passed.')
